# devpick-api-spec.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Trace API",
    "description": "Trace 백엔드 REST API — 개발자 성장형 통합 플랫폼\n\n**Base URL**: `https://api.devpick.kr/v1`\n\n**인증**: `Authorization: Bearer {access_token}` (JWT)",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "로컬 개발 서버"
    },
    {
      "url": "https://api.devpick.kr/v1",
      "description": "프로덕션 서버"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "ApiResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" },
          "data": {},
          "error": {
            "type": "object",
            "nullable": true,
            "properties": {
              "code": { "type": "string" },
              "message": { "type": "string" },
              "detail": {}
            }
          }
        }
      },
...
      "Level": {
        "type": "string",
        "enum": ["BEGINNER", "JUNIOR", "MIDDLE", "SENIOR"]
      },
...
      "PostCreateRequest": {
        "type": "object",
        "required": ["title", "content", "level"],
        "properties": {
          "title": { "type": "string", "maxLength": 500, "example": "Spring Boot에서 JPA N+1 문제 해결 방법이 궁금합니다" },
          "content": { "type": "string", "example": "Fetch Join을 사용해야 하는지..." },
          "level": { "$ref": "#/components/schemas/Level" }
        }
      },
      "PostUpdateRequest": {
        "type": "object",
        "required": ["title", "content", "level"],
        "properties": {
          "title": { "type": "string", "maxLength": 500 },
          "content": { "type": "string" },
          "level": { "$ref": "#/components/schemas/Level" }
        }
      },
      "PostSummaryResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "title": { "type": "string" },
          "level": { "$ref": "#/components/schemas/Level" },
          "authorId": { "type": "string", "format": "uuid" },
          "authorNickname": { "type": "string" },
          "answerCount": { "type": "integer", "format": "int64" },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
      "PostDetailResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "title": { "type": "string" },
          "content": { "type": "string" },
          "level": { "$ref": "#/components/schemas/Level" },
          "authorId": { "type": "string", "format": "uuid" },
          "authorNickname": { "type": "string" },
          "answerCount": { "type": "integer", "format": "int64" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "PostListResponse": {
        "type": "object",
        "properties": {
          "posts": { "type": "array", "items": { "$ref": "#/components/schemas/PostSummaryResponse" } },
          "page": { "type": "integer" },
          "size": { "type": "integer" },
          "totalElements": { "type": "integer", "format": "int64" },
          "totalPages": { "type": "integer" }
        }
      },
...
      "ContentInfo": {
        "type": "object",
        "nullable": true,
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "title": { "type": "string" },
          "preview": { "type": "string" }
        }
      },
      "PostInfo": {
        "type": "object",
        "nullable": true,
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "title": { "type": "string" }
        }
      },
      "HistoryItemResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "actionType": {
            "type": "string",
            "description": "content_opened, ai_summary_viewed, scrapped, question_created, post_created, content_liked, ai_quiz_completed 등"
          },
          "points": { "type": "integer", "nullable": true, "description": "해당 행동으로 적립된 포인트(해당 시)" },
          "content": { "$ref": "#/components/schemas/ContentInfo" },
          "post": { "$ref": "#/components/schemas/PostInfo" },
          "answer": {
            "type": "object",
            "nullable": true,
            "properties": { "id": { "type": "string", "format": "uuid" } }
          },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
      "HistoryPageResponse": {
        "type": "object",
        "properties": {
          "items": { "type": "array", "items": { "$ref": "#/components/schemas/HistoryItemResponse" } },
          "page": { "type": "integer" },
          "size": { "type": "integer" },
          "totalElements": { "type": "integer", "format": "int64" },
          "totalPages": { "type": "integer" }
        }
      },
      "ActivityItemResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "actionType": {
            "type": "string",
            "enum": ["content_opened", "ai_summary_viewed", "scrapped", "question_created", "post_created", "content_liked"]
          },
          "content": { "$ref": "#/components/schemas/ContentInfo" },
          "post": { "$ref": "#/components/schemas/PostInfo" },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
      "ActivityPageResponse": {
        "type": "object",
        "properties": {
          "items": { "type": "array", "items": { "$ref": "#/components/schemas/ActivityItemResponse" } },
          "page": { "type": "integer" },
          "size": { "type": "integer" },
          "totalElements": { "type": "integer", "format": "int64" },
          "totalPages": { "type": "integer" }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": false },
          "error": {
            "type": "object",
            "properties": {
              "code": { "type": "string", "example": "AUTH_001" },
              "message": { "type": "string" },
              "detail": {}
            }
          }
        }
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "인증 필요",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": { "success": false, "error": { "code": "AUTH_001", "message": "로그인이 필요합니다" } }
          }
        }
      },
      "NotFound": {
        "description": "리소스 없음",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      }
    }
  },
  "security": [],
  "paths": {
...
    "/posts": {
      "get": {
        "tags": ["Community"],
        "summary": "게시글 목록 조회",
        "description": "query가 있으면 제목·본문 부분 일치 검색합니다.",
        "parameters": [
          { "name": "query", "in": "query", "required": false, "schema": { "type": "string" }, "description": "검색 키워드" },
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 0 } },
          { "name": "size", "in": "query", "schema": { "type": "integer", "default": 20 } }
        ],
        "responses": {
          "200": {
            "description": "게시글 목록",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ApiResponse" },
                    { "properties": { "data": { "$ref": "#/components/schemas/PostListResponse" } } }
                  ]
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Community"],
        "summary": "질문 작성",
        "description": "question_created 히스토리가 기록됩니다.",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/PostCreateRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "작성 성공",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ApiResponse" },
                    { "properties": { "data": { "$ref": "#/components/schemas/PostDetailResponse" } } }
                  ]
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
...
    "/posts/{postId}": {
      "get": {
        "tags": ["Community"],
        "summary": "게시글 상세 조회",
        "parameters": [
          { "name": "postId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": {
            "description": "조회 성공",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ApiResponse" },
                    { "properties": { "data": { "$ref": "#/components/schemas/PostDetailResponse" } } }
                  ]
                }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "put": {
        "tags": ["Community"],
        "summary": "게시글 수정",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "postId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/PostUpdateRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "수정 성공",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ApiResponse" },
                    { "properties": { "data": { "$ref": "#/components/schemas/PostDetailResponse" } } }
                  ]
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "description": "작성자 아님" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "delete": {
        "tags": ["Community"],
        "summary": "게시글 삭제",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "postId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "204": { "description": "삭제 성공" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "description": "작성자 아님" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
    "/posts/{postId}/like": {
      "post": {
        "tags": ["Community"],
        "summary": "게시글 좋아요",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "postId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "201": { "description": "성공" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "description": "이미 좋아요함" }
        }
      },
      "delete": {
        "tags": ["Community"],
        "summary": "게시글 좋아요 취소",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "postId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "204": { "description": "취소 성공" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "description": "좋아요 없음" }
        }
      }
    },
...
    "/history": {
      "get": {
        "tags": ["History"],
        "summary": "학습 히스토리 조회",
        "description": "content_liked를 제외한 학습 행동 기록을 최신순으로 반환합니다. actionTypes·기간 필터를 지원합니다.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "actionTypes",
            "in": "query",
            "required": false,
            "schema": { "type": "array", "items": { "type": "string" } },
            "style": "form",
            "explode": true,
            "description": "포함할 actionType 목록 (미입력 시 전체, 단 content_liked는 제외)"
          },
          { "name": "startDate", "in": "query", "required": false, "schema": { "type": "string", "format": "date-time" } },
          { "name": "endDate", "in": "query", "required": false, "schema": { "type": "string", "format": "date-time" } },
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 0, "minimum": 0 } },
          { "name": "size", "in": "query", "schema": { "type": "integer", "default": 20, "minimum": 1, "maximum": 100 } }
        ],
        "responses": {
          "200": {
            "description": "히스토리 조회 성공",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ApiResponse" },
                    { "properties": { "data": { "$ref": "#/components/schemas/HistoryPageResponse" } } }
                  ]
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/history/activity": {
      "get": {
        "tags": ["History"],
        "summary": "전체 활동 내역 조회",
        "description": "content_liked 포함 모든 활동 내역을 최신순으로 반환합니다.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 0, "minimum": 0 } },
          { "name": "size", "in": "query", "schema": { "type": "integer", "default": 20, "minimum": 1, "maximum": 100 } }
        ],
        "responses": {
          "200": {
            "description": "활동 내역 조회 성공",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ApiResponse" },
                    { "properties": { "data": { "$ref": "#/components/schemas/ActivityPageResponse" } } }
                  ]
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
...
  },
...
}